Let env variable override the config flag-based VVP_EXECUTABLE
This is much more convenient for use in bazel via Bluespec.  The value
is set on the outside and the location of the build directory no
longer matters. This value was previously wired into vvp.conf which is
populated with a path containing --prefix. Naturally this is not at
all the place where vvp is in bazel builds.